+++ mirtop/gff/merge.py
@@ -30,8 +30,8 @@ def _fix(line, expression):
     return paste_columns(cols, guess_format(line))
 
 def _convert_to_string(d, s):
-    v = [str(d[ss]) if ss in d else 0 for ss in s]
-    return " Expression %s" % ",".join(v)
+    v = [str(d[ss]) if ss in d else "0" for ss in s]
+    return "%s" % ",".join(v)
 
 def _chrom(string):
     return string.split("\t")[0]