+++ HISTORY.md
@@ -1,6 +1,9 @@
 - 0.3.*a
 
- * Fix bug in order of merged gff file. Thanks AlisR.
+ * Fix Expression attrb when join gff files.Thanks @AlisR.
+ * Print help when no files are giving to any subcommand.
+ * Fix bug for duplicated isomiRs tags. Thanks @AlisR.
+ * Fix bug in order of merged gff file. Thanks @AlisR.
  * Add module to read GFF/GTF line in body.py
  * Add version line to stats output
  * Improve PROST! importer